requirements.txt
Django
django-ninja
orjson
psycopg2-binary
reactpy